requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
deprecation==2.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
postgrest==1.1.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
realtime==2.7.0
requests==2.32.5
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
storage3==0.12.1
strenum==0.4.15
supabase==2.18.1
supabase-auth==2.12.3
supabase-functions==0.10.1
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
